+++ test/FamaProfiling_server_test.py
@@ -106,12 +106,13 @@ class FamaProfilingTest(unittest.TestCase):
 #    @unittest.skip("")
     def test_pe_fama_profiling(self):
 
-        ret = self.getImpl().run_FamaProfiling(self.getContext(), {'workspace_name': self.ws_info[1],
+        ret = self.getImpl().run_FamaReadProfiling(self.getContext(), {'workspace_name': self.ws_info[1],
                                                                     'ref_dataset': 'nitrogen',
+                                                                    'is_paired_end': 1,
 #                                                                    'read_library_ref': '22763/10/1',  # 4706 read set (LARGE!)
 #                                                                   'read_library_ref': '22763/20/1',  # 4701 2M susbset
 #                                                                    'read_library_ref': '22763/2/1',  # tiny read set (20 + 20 reads)
-                                                                    'read_library_ref': self.pe_reads_ref['obj_ref'], # tiny read set (20 + 20 reads), upload from local files
+                                                                    'read_library_refs': [self.pe_reads_ref['obj_ref'], '22763/2/1'], # tiny read set (20 + 20 reads), upload from local files
                                                                     'output_read_library_name': 'Fama_pe_test_output'})
         print ('Report name', ret[0]['report_name'])
         print ('Report reference', ret[0]['report_ref'])
@@ -119,10 +120,11 @@ class FamaProfilingTest(unittest.TestCase):
         
 #    @unittest.skip("")
     def test_se_fama_profiling(self):
-        ret = self.getImpl().run_FamaProfiling(self.getContext(), {'workspace_name': self.ws_info[1],
+        ret = self.getImpl().run_FamaReadProfiling(self.getContext(), {'workspace_name': self.ws_info[1],
                                                                     'ref_dataset': 'nitrogen',
+                                                                    'is_paired_end': 0,
 #                                                                    'read_library_ref': '22763/33/1',  # tiny read set (20 reads)
-                                                                    'read_library_ref': self.se_reads_ref['obj_ref'],
+                                                                    'read_library_refs': [self.se_reads_ref['obj_ref'], '22763/33/1'],
                                                                     'output_read_library_name': 'Fama_se_test_output'})
         print ('Report name', ret[0]['report_name'])
         print ('Report reference', ret[0]['report_ref'])
@@ -130,7 +132,7 @@ class FamaProfilingTest(unittest.TestCase):
         
 #    @unittest.skip("")
     def test_protein_fama_profiling(self):
-        ret = self.getImpl().run_FamaProteinProfiling(self.getContext(), {'workspace_name': self.ws_info[1],
+        ret = self.getImpl().run_FamaGenomeProfiling(self.getContext(), {'workspace_name': self.ws_info[1],
                                                                     'ref_dataset': 'nitrogen',
                                                                     'genome_ref': ['22763/32/1', '41747/12/1'],  # S. oneidensis and S. amazonensis genomes
                                                                     'output_annotation_name': 'Fama_Ncycle.',